// ler/src/enum/beOne.enum.ts
export enum BeOneOrderStatus {
  PICKED_UP_COMPLETED = 'Pickup Completed /Scanned',
  RECEIVED_AT_WAREHOUSE = 'Received at Warehouse',
  READY_TO_DISPATCH = 'Ready to Dispatch',
  SHELVED = 'Shelved',
  OFD_DISPATCHED = 'OFD/Dispatched',
  RESCHEDULED_FOR_NEXT_DAY = 'Rescheduled for Next Day Delivery',
  IN_TRANSIT = 'In Transit',
  ITEM_DELIVERED = 'Delivered'
}