// ler/src/modules/grpc/proto/ler.pb.ts
export interface CreatePickupRequest {
  referenceNo: string;
  originCity: string;
  destinationCity: string;
  senderName: string;
  senderPhone: string;
  senderAddress: string;
  receiverName: string;
  receiverPhone: string;
  receiverAddress: string;
  trackingNumber: string;
  description: string;
}
// ...
export interface CreatePickupForAccessoriesRequest {
  referenceNo: string;
  originCity: string;
  destinationCity: string;
  senderName: string;
  senderPhone: string;
  senderAddress: string;
  receiverName: string;
  receiverPhone: string;
  receiverAddress: string;
  trackingNumber: string;
  description: string;
  skudetails: SkuDetails[];
}

export interface CreatePickupResponse {
  awbNo: string;
}

export interface CreateShipmentResponse {
  trackingNumber: string;
}

export interface GetPickupStatusRequest {
  awbNo: string[];
  isDelivered: boolean;
}

export interface SkuDetails {
  sku: string;
  description: string;
  cod: string;
  piece: string;
  weight: string;
}